add ObterPerfilUsuario

+++ src/VilevePay.Application/ViewModels/v1/Parametrizacao/PerfilUsuarioViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class PerfilUsuarioViewModel
     {
-        public int CodigoPerfil { get; set; }
+        public long CodigoPerfil { get; set; }
         public string Nome { get; set; }
     }
 }
